Preparatory: Add Kenya to question and response list
This commit adds Kenya to question and response list, this triggers
this generation of regression test artefacts for Kenya. This sets the
basis for evaluating the content changes in the commits to come.

These changes only affect all opposite and same sex outcomes.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.