Regenerate artefacts for LandLord Immigration Check
Add British or Irish option

This commit shows the question and outcomes after including the
british-or-irish option to the regression test configuration files.

This addition triggers the creation of british-or-irish regression artefacts.

This gives assurance that this setup will form the basis for further future
changes in later commits.

See these documents [1] and [2] for more details

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned commit/changes have had no
(interfering) effect to other smart answers.

[1] https://docs.google.com/document/d/1PtZxv9Fy-SOUtRSDvUq7_jbqLYUBn8Pp0nCf57U2UIE/edit
[2] https://docs.google.com/drawings/d/1x2547adffe616JMZ6qnI2ST5H2f-OnnUwSZ5isn5dzM/edit

NB:
The following nodes are not being excerised and hence failing at this point
 - :has_asylum_card?
 - :has_certificate?
 - :has_documents?
 - :has_residence_card_or_eu_eea_swiss_family_member?
 - :has_uk_passport?
 - :immigration_application?
 - :outcome_can_rent_but_check_will_be_needed_again
 - :outcome_can_rent_for_12_months
 - :outcome_landlords_checking_service
 - :right_to_abode?
 - :tenant_country?
 - :time_limited_to_remain?
 - :waiting_for_documents?