Implement British-or-Irish route/ flow
This commit updates the LandlordImmigrationCheckFlow by adding logic
that will direct users to the corresponding question or outcomes based on
their response to questions asked.

The british-or-irish option has also been added to the erb template to ensure
that it is displayed to users

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned commit/changes have had no
(interfering) effect to other smart answers.

See these documents [1] and [2] for more details

[1] https://docs.google.com/document/d/1PtZxv9Fy-SOUtRSDvUq7_jbqLYUBn8Pp0nCf57U2UIE/edit
[2] https://docs.google.com/drawings/d/1x2547adffe616JMZ6qnI2ST5H2f-OnnUwSZ5isn5dzM/edit

NB:
The following nodes are not being excerised and hence failing at this point
 - :has_asylum_card?
 - :has_certificate?
 - :has_documents?
 - :has_residence_card_or_eu_eea_swiss_family_member?
 - :has_uk_passport?
 - :immigration_application?
 - :outcome_can_rent_but_check_will_be_needed_again
 - :outcome_can_rent_for_12_months
 - :outcome_landlords_checking_service
 - :right_to_abode?
 - :tenant_country?
 - :time_limited_to_remain?
 - :waiting_for_documents?